Added Weblis batch search tools
Tools are not finished but they can be used. They are not finished
because I don't know what I want them to become yet. Right now they are
set up to search Webflis and write html files from source pages, and the
text file must have the nsn and what to search for on the page.